Create Page API

+++ client_admin/src/helpers/axios.js
@@ -23,8 +23,8 @@ axiosIntance.interceptors.request.use((req) => {
 axiosIntance.interceptors.response.use((res) => {
     return res;
 }, (error) => {
-    const { status } = error.response;
-    if (status === 500) {
+    const status = error.response ? error.response.status : 500;
+    if (status && status === 500) {
         localStorage.clear();
         store.dispatch({ type: authConstants.LOGOUT_SUCCESS });
     }
